+++ tests/testthat/test_first_of.R
@@ -5,16 +5,33 @@ context("first_of")
 # year --------------------------------------------------------------------
 
 test_that("first_of_year works as expected", {
-
-  x <- as.Date("2018-01-03")
-
+  # for Date
   expect_identical(
-    first_of_year(x),
+    first_of_year(as.Date("2018-01-03")),
     as.Date("2018-01-01")
   )
-
   expect_identical(
-    last_of_year(x),
+    last_of_year(as.Date("2018-01-03")),
+    as.Date("2018-12-31")
+  )
+
+  # for date_xx
+  expect_identical(
+    first_of_year(date_yq(2018, 2)),
+    as.Date("2018-01-01")
+  )
+  expect_identical(
+    last_of_year(date_ym(2018, 1)),
+    as.Date("2018-12-31")
+  )
+
+  # for numeric years
+  expect_identical(
+    first_of_year(2018),
+    as.Date("2018-01-01")
+  )
+  expect_identical(
+    last_of_year(2018),
     as.Date("2018-12-31")
   )
 })
@@ -22,7 +39,6 @@ test_that("first_of_year works as expected", {
 
 
 test_that("test first_of_month against lubridate", {
-
   if (!requireNamespace("lubridate", quietly = TRUE)){
     skip("test requires lubridate")
   }
@@ -41,12 +57,10 @@ test_that("test first_of_month against lubridate", {
 })
 
 
+
+
 # quarter -----------------------------------------------------------------
-
-
-
 test_that("first_of_quarter works as expected", {
-
   expect_identical(
     first_of_quarter(make_date(2000, 1:12, 1)),
     c(
@@ -56,7 +70,6 @@ test_that("first_of_quarter works as expected", {
       rep(make_date(2000, 10, 1), 3)
     )
   )
-
 
   expect_identical(
     last_of_quarter(make_date(2000, 1:12, 1)),
@@ -68,13 +81,23 @@ test_that("first_of_quarter works as expected", {
     )
   )
 
+  expect_identical(
+    unique(first_of_quarter(make_date(2000, 1:12, 1))),
+    first_of_yq(2000, 1:4)
+  )
+  expect_identical(
+    last_of_quarter(unique(last_of_quarter(make_date(2000, 1:12, 1)))),
+    last_of_yq(2000, 1:4)
+  )
+
+  expect_identical(first_of_yq("2004-05-07"), as.Date("2004-04-01"))
+  expect_identical(last_of_yq("2004-05-07"), as.Date("2004-06-30"))
 })
 
 
 
 
 test_that("test first_of_quarter against lubridate", {
-
   if (!requireNamespace("lubridate", quietly = TRUE)){
     skip("test requires lubridate")
   }
@@ -85,7 +108,6 @@ test_that("test first_of_quarter against lubridate", {
     first_of_quarter(tdat),
     lubridate::floor_date(tdat, "quarter")
   )
-
   expect_identical(
     last_of_quarter(tdat),
     lubridate::ceiling_date(tdat, "quarter") - 1
@@ -98,7 +120,6 @@ test_that("test first_of_quarter against lubridate", {
 # month -------------------------------------------------------------------
 
 test_that("first_of_month works as expected", {
-
   # first day
   expect_identical(
     first_of_ym(2018, c(4, 12)),
@@ -116,7 +137,6 @@ test_that("first_of_month works as expected", {
       as.Date(c("2000-02-29", "2004-02-29", "2100-02-28"))
     )
 
-
   expect_identical(
     last_of_month(as.Date("2018-05-13")),
     as.Date("2018-05-31")
@@ -132,7 +152,6 @@ test_that("first_of_month works as expected", {
 
 
 test_that("test first_of_month against lubridate", {
-
   if (!requireNamespace("lubridate", quietly = TRUE)){
     skip("test requires lubridate")
   }
@@ -157,7 +176,6 @@ test_that("test first_of_month against lubridate", {
 
 
 test_that("first_of_isoweek works", {
-
   tdat <- seq(as.Date("2004-12-20"), as.Date("2020-01-06"), by = "7 days")
 
   expect_identical(
@@ -169,6 +187,18 @@ test_that("first_of_isoweek works", {
     last_of_isoweek(tdat),
     tdat + 6L
   )
+
+  expect_identical(
+    seq(as.Date("2004-01-05"), as.Date("2004-12-20"), by = "7 days"),
+    first_of_yw(2004, 2:52)
+  )
+  expect_identical(
+    seq(as.Date("2004-01-11"), as.Date("2004-12-26"), by = "7 days"),
+    last_of_yw(2004, 2:52)
+  )
+
+  expect_identical(first_of_yw("2004-01-07"), as.Date("2004-01-05"))
+  expect_identical(last_of_yw("2004-01-07"), as.Date("2004-01-11"))
 })
 
 
@@ -180,7 +210,6 @@ test_that("first_of_isoyear works", {
     from = as.Date(c("2018-01-01", "2017-01-02", "2016-01-04", "2005-01-03", "2003-12-29")),
     to   = as.Date(c("2018-12-30", "2017-12-31", "2017-01-01", "2006-01-01", "2005-01-02"))
   )
-
 
   expect_identical(
     first_of_isoyear(as_date_y(eres$year)),